Improved Grammar and Clarity of What is Webpack
I've gone through and resolved many of the minor grammatical issues in this section, including sentence flow (run on sentences, garden path sentences), ambiguous pronouns, and clarity of intent. As a result, the chapter flows much better and avoids distracting the reader.

I've maintained the author's original intent as much as possible. Thanks :)